Pull

+++ app/Http/Controllers/ForumController.php
@@ -14,10 +14,9 @@
 namespace App\Http\Controllers;
 
 use App\Models\Post;
-use App\Models\User;
 use App\Models\Forum;
 use App\Models\Topic;
-use Brian2694\Toastr\Toastr;
+use Illuminate\Support\Str;
 use Illuminate\Http\Request;
 use App\Repositories\ChatRepository;
 use App\Achievements\UserMade25Posts;
@@ -47,22 +46,15 @@ class ForumController extends Controller
     private $chat;
 
     /**
-     * @var Toastr
-     */
-    private $toastr;
-
-    /**
      * ForumController Constructor.
      *
      * @param TaggedUserRepository $tag
      * @param ChatRepository       $chat
-     * @param Toastr               $toastr
      */
-    public function __construct(TaggedUserRepository $tag, ChatRepository $chat, Toastr $toastr)
+    public function __construct(TaggedUserRepository $tag, ChatRepository $chat)
     {
         $this->tag = $tag;
         $this->chat = $chat;
-        $this->toastr = $toastr;
     }
 
     /**
@@ -76,7 +68,7 @@ class ForumController extends Controller
     {
         $categories = Forum::oldest('position')->get();
 
-        $user = auth()->user();
+        $user = $request->user();
 
         $pests = $user->group->permissions->where('show_forum', '=', 0)->pluck('forum_id')->toArray();
         if (! is_array($pests)) {
@@ -157,24 +149,25 @@ class ForumController extends Controller
                 }
             }
         }
-        $direction = 2;
-        $order = 'desc';
-        if ($request->has('direction') && $request->input('direction') == 1) {
-            $direction = 1;
-            $order = 'asc';
-        }
+
         if ($request->has('body') && $request->input('body') != '') {
-            $sorting = 'posts.id';
-            if ($request->has('sorting') && $request->input('sorting') == 'created_at') {
-                $sorting = 'posts.created_at';
+            if ($request->has('sorting') && $request->input('sorting') != null) {
+                $sorting = "posts.{$request->input('sorting')}";
+                $direction = $request->input('direction');
+            } else {
+                $sorting = 'posts.id';
+                $direction = 'desc';
             }
             $results = $result->orderBy($sorting, $direction)->paginate(25);
         } else {
-            $sorting = 'topics.last_reply_at';
-            if ($request->has('sorting') && $request->input('sorting') == 'created_at') {
-                $sorting = 'topics.created_at';
+            if ($request->has('sorting') && $request->input('sorting') != null) {
+                $sorting = "topics.{$request->input('sorting')}";
+                $direction = $request->input('direction');
+            } else {
+                $sorting = 'topics.last_reply_at';
+                $direction = 'desc';
             }
-            $results = $result->orderBy($sorting, $order)->paginate(25);
+            $results = $result->orderBy($sorting, $direction)->paginate(25);
         }
 
         $results->setPath('?name='.$request->input('name'));
@@ -210,7 +203,7 @@ class ForumController extends Controller
      */
     public function subscriptions(Request $request)
     {
-        $user = auth()->user();
+        $user = $request->user();
 
         $pests = $user->group->permissions->where('show_forum', '=', 0)->pluck('forum_id')->toArray();
         if (! is_array($pests)) {
@@ -267,7 +260,7 @@ class ForumController extends Controller
      */
     public function latestTopics(Request $request)
     {
-        $user = auth()->user();
+        $user = $request->user();
 
         $pests = $user->group->permissions->where('show_forum', '=', 0)->pluck('forum_id')->toArray();
         if (! is_array($pests)) {
@@ -301,7 +294,7 @@ class ForumController extends Controller
      */
     public function latestPosts(Request $request)
     {
-        $user = auth()->user();
+        $user = $request->user();
 
         $pests = $user->group->permissions->where('show_forum', '=', 0)->pluck('forum_id')->toArray();
         if (! is_array($pests)) {
@@ -379,7 +372,7 @@ class ForumController extends Controller
         $category = Forum::findOrFail($forum->id);
         if ($category->getPermission()->show_forum != true) {
             return redirect()->route('forum_index')
-                ->with($this->toastr->error('You Do Not Have Access To This Category!', 'Whoops!', ['options']));
+                ->withErrors('You Do Not Have Access To This Category!');
         }
 
         // Fetch topics->posts in descending order
@@ -424,7 +417,7 @@ class ForumController extends Controller
         $category = Forum::findOrFail($forum->parent_id);
         if ($category->getPermission()->show_forum != true) {
             return redirect()->route('forum_index')
-                ->with($this->toastr->error('You Do Not Have Access To This Forum!', 'Whoops!', ['options']));
+                ->withErrors('You Do Not Have Access To This Forum!');
         }
 
         // Fetch topics->posts in descending order
@@ -469,7 +462,7 @@ class ForumController extends Controller
         if ($category->getPermission()->read_topic != true) {
             // Redirect him to the forum index
             return redirect()->route('forum_index')
-                ->with($this->toastr->error('You Do Not Have Access To Read This Topic!', 'Whoops!', ['options']));
+                ->withErrors('You Do Not Have Access To Read This Topic!');
         }
 
         // Increment view
@@ -496,15 +489,15 @@ class ForumController extends Controller
      */
     public function reply(Request $request, $slug, $id)
     {
-        $user = auth()->user();
+        $user = $request->user();
         $topic = Topic::findOrFail($id);
         $forum = $topic->forum;
         $category = $forum->getCategory();
 
         // The user has the right to create a topic here?
-        if (! $category->getPermission()->reply_topic || ($topic->state == 'close' && ! auth()->user()->group->is_modo)) {
+        if (! $category->getPermission()->reply_topic || ($topic->state == 'close' && ! $request->user()->group->is_modo)) {
             return redirect()->route('forum_index')
-                ->with($this->toastr->error('You Cannot Reply To This Topic!', 'Whoops!', ['options']));
+                ->withErrors('You Cannot Reply To This Topic!');
         }
 
         $post = new Post();
@@ -520,7 +513,7 @@ class ForumController extends Controller
 
         if ($v->fails()) {
             return redirect()->route('forum_topic', ['slug' => $topic->slug, 'id' => $topic->id])
-                ->with($this->toastr->error($v->errors()->toJson(), 'Whoops!', ['options']));
+                ->withErrors($v->errors());
         } else {
             $post->save();
 
@@ -609,8 +602,8 @@ class ForumController extends Controller
             $user->addProgress(new UserMade800Posts(), 1);
             $user->addProgress(new UserMade900Posts(), 1);
 
-            return redirect($realUrl)
-                ->with($this->toastr->success('Post Successfully Posted', 'Yay!', ['options']));
+            return redirect()->to($realUrl)
+                ->withSuccess('Post Successfully Posted');
         }
     }
 
@@ -631,7 +624,7 @@ class ForumController extends Controller
         // The user has the right to create a topic here?
         if ($category->getPermission()->start_topic != true) {
             return redirect()->route('forum_index')
-                ->with($this->toastr->error('You Cannot Start A New Topic Here!', 'Whoops!', ['options']));
+                ->withErrors('You Cannot Start A New Topic Here!');
         }
 
         return view('forum.new_topic', [
@@ -652,20 +645,20 @@ class ForumController extends Controller
      */
     public function newTopic(Request $request, $slug, $id)
     {
-        $user = auth()->user();
+        $user = $request->user();
         $forum = Forum::findOrFail($id);
         $category = $forum->getCategory();
 
         // The user has the right to create a topic here?
         if ($category->getPermission()->start_topic != true) {
             return redirect()->route('forum_index')
-                ->with($this->toastr->error('You Cannot Start A New Topic Here!', 'Whoops!', ['options']));
+                ->withErrors('You Cannot Start A New Topic Here!');
         }
 
         // Create The Topic
         $topic = new Topic();
         $topic->name = $request->input('title');
-        $topic->slug = str_slug($request->input('title'));
+        $topic->slug = Str::slug($request->input('title'));
         $topic->state = 'open';
         $topic->first_post_user_id = $topic->last_post_user_id = $user->id;
         $topic->first_post_user_username = $topic->last_post_user_username = $user->username;
@@ -689,7 +682,7 @@ class ForumController extends Controller
 
         if ($v->fails()) {
             return redirect()->route('forum_index')
-                ->with($this->toastr->error($v->errors()->toJson(), 'Whoops!', ['options']));
+                ->withErrors($v->errors());
         } else {
             $topic->save();
 
@@ -706,7 +699,7 @@ class ForumController extends Controller
 
             if ($v->fails()) {
                 return redirect()->route('forum_index')
-                    ->with($this->toastr->error($v->errors()->toJson(), 'Whoops!', ['options']));
+                    ->withErrors($v->errors());
             } else {
                 $post->save();
                 $topic->num_post = 1;
@@ -745,7 +738,7 @@ class ForumController extends Controller
                 $user->addProgress(new UserMade900Posts(), 1);
 
                 return redirect()->route('forum_topic', ['slug' => $topic->slug, 'id' => $topic->id])
-                    ->with($this->toastr->success('Topic Created Successfully!', 'Yay!', ['options']));
+                    ->withSuccess('Topic Created Successfully!');
             }
         }
     }
@@ -777,10 +770,10 @@ class ForumController extends Controller
      */
     public function editTopic(Request $request, $slug, $id)
     {
-        $user = auth()->user();
+        $user = $request->user();
         $topic = Topic::findOrFail($id);
 
-        abort_unless($user->group->is_modo, 403);
+        abort_unless($user->group->is_modo || $user->id === $topic->first_post_user_id, 403);
         $name = $request->input('name');
         $forum_id = $request->input('forum_id');
         $topic->name = $name;
@@ -788,7 +781,7 @@ class ForumController extends Controller
         $topic->save();
 
         return redirect()->route('forum_topic', ['slug' => $topic->slug, 'id' => $topic->id])
-            ->with($this->toastr->success('Topic Successfully Edited', 'Yay!', ['options']));
+            ->withSuccess('Topic Successfully Edited');
     }
 
     /**
@@ -825,16 +818,16 @@ class ForumController extends Controller
      */
     public function postEdit(Request $request, $postId)
     {
-        $user = auth()->user();
+        $user = $request->user();
         $post = Post::findOrFail($postId);
         $postUrl = "forums/topic/{$post->topic->slug}.{$post->topic->id}?page={$post->getPageNumber()}#post-{$postId}";
 
-        abort_unless($user->group->is_modo || $post->user_id == $user->id, 403);
+        abort_unless($user->group->is_modo || $user->id === $post->user_id, 403);
         $post->content = $request->input('content');
         $post->save();
 
-        return redirect($postUrl)
-            ->with($this->toastr->success('Post Successfully Edited!', 'Yay!', ['options']));
+        return redirect()->to($postUrl)
+            ->withSuccess('Post Successfully Edited!');
     }
 
     /**
@@ -844,16 +837,16 @@ class ForumController extends Controller
      *
      * @return Illuminate\Http\RedirectResponse
      */
-    public function postDelete($postId)
+    public function postDelete(Request $request, $postId)
     {
-        $user = auth()->user();
+        $user = $request->user();
         $post = Post::with('topic')->findOrFail($postId);
 
-        abort_unless($user->group->is_modo || $post->user_id == $user->id, 403);
+        abort_unless($user->group->is_modo || $user->id === $post->user_id, 403);
         $post->delete();
 
         return redirect()->route('forum_topic', ['slug' => $post->topic->slug, 'id' => $post->topic->id])
-            ->with($this->toastr->success('This Post Is Now Deleted!', 'Success', ['options']));
+            ->withSuccess('This Post Is Now Deleted!');
     }
 
     /**
@@ -864,9 +857,9 @@ class ForumController extends Controller
      *
      * @return Illuminate\Http\RedirectResponse
      */
-    public function closeTopic($slug, $id)
+    public function closeTopic(Request $request, $slug, $id)
     {
-        $user = auth()->user();
+        $user = $request->user();
         $topic = Topic::findOrFail($id);
 
         abort_unless($user->group->is_modo || $user->id === $topic->first_post_user_id, 403);
@@ -874,7 +867,7 @@ class ForumController extends Controller
         $topic->save();
 
         return redirect()->route('forum_topic', ['slug' => $topic->slug, 'id' => $topic->id])
-            ->with($this->toastr->success('This Topic Is Now Closed!', 'Success', ['options']));
+            ->withSuccess('This Topic Is Now Closed!');
     }
 
     /**
@@ -885,9 +878,9 @@ class ForumController extends Controller
      *
      * @return Illuminate\Http\RedirectResponse
      */
-    public function openTopic($slug, $id)
+    public function openTopic(Request $request, $slug, $id)
     {
-        $user = auth()->user();
+        $user = $request->user();
         $topic = Topic::findOrFail($id);
 
         abort_unless($user->group->is_modo || $user->id === $topic->first_post_user_id, 403);
@@ -895,7 +888,7 @@ class ForumController extends Controller
         $topic->save();
 
         return redirect()->route('forum_topic', ['slug' => $topic->slug, 'id' => $topic->id])
-            ->with($this->toastr->success('This Topic Is Now Open!', 'Success', ['options']));
+            ->withSuccess('This Topic Is Now Open!');
     }
 
     /**
@@ -906,18 +899,18 @@ class ForumController extends Controller
      *
      * @return Illuminate\Http\RedirectResponse
      */
-    public function deleteTopic($slug, $id)
+    public function deleteTopic(Request $request, $slug, $id)
     {
-        $user = auth()->user();
+        $user = $request->user();
         $topic = Topic::findOrFail($id);
 
-        abort_unless($user->group->is_modo, 403);
+        abort_unless($user->group->is_modo || $user->id === $topic->first_post_user_id, 403);
         $posts = $topic->posts();
         $posts->delete();
         $topic->delete();
 
         return redirect()->route('forum_display', ['slug' => $topic->forum->slug, 'id' => $topic->forum->id])
-            ->with($this->toastr->error('This Topic Is Now Deleted!', 'Warning', ['options']));
+            ->withErrors('This Topic Is Now Deleted!');
     }
 
     /**
@@ -935,7 +928,7 @@ class ForumController extends Controller
         $topic->save();
 
         return redirect()->route('forum_topic', ['slug' => $topic->slug, 'id' => $topic->id])
-            ->with($this->toastr->success('This Topic Is Now Pinned!', 'Success', ['options']));
+            ->withSuccess('This Topic Is Now Pinned!');
     }
 
     /**
@@ -953,7 +946,7 @@ class ForumController extends Controller
         $topic->save();
 
         return redirect()->route('forum_topic', ['slug' => $topic->slug, 'id' => $topic->id])
-            ->with($this->toastr->success('This Topic Is Now Unpinned!', 'Success', ['options']));
+            ->withSuccess('This Topic Is Now Unpinned!');
     }
 
     /**
@@ -975,7 +968,7 @@ class ForumController extends Controller
         $topic->save();
 
         return redirect()->route('forum_topic', ['slug' => $topic->slug, 'id' => $topic->id])
-            ->with($this->toastr->info('Label Change Has Been Applied', 'Info', ['options']));
+            ->withInfo('Label Change Has Been Applied');
     }
 
     public function deniedTopic($slug, $id)
@@ -989,7 +982,7 @@ class ForumController extends Controller
         $topic->save();
 
         return redirect()->route('forum_topic', ['slug' => $topic->slug, 'id' => $topic->id])
-            ->with($this->toastr->info('Label Change Has Been Applied', 'Info', ['options']));
+            ->withInfo('Label Change Has Been Applied');
     }
 
     public function solvedTopic($slug, $id)
@@ -1003,7 +996,7 @@ class ForumController extends Controller
         $topic->save();
 
         return redirect()->route('forum_topic', ['slug' => $topic->slug, 'id' => $topic->id])
-            ->with($this->toastr->info('Label Change Has Been Applied', 'Info', ['options']));
+            ->withInfo('Label Change Has Been Applied');
     }
 
     public function invalidTopic($slug, $id)
@@ -1017,7 +1010,7 @@ class ForumController extends Controller
         $topic->save();
 
         return redirect()->route('forum_topic', ['slug' => $topic->slug, 'id' => $topic->id])
-            ->with($this->toastr->info('Label Change Has Been Applied', 'Info', ['options']));
+            ->withInfo('Label Change Has Been Applied');
     }
 
     public function bugTopic($slug, $id)
@@ -1031,7 +1024,7 @@ class ForumController extends Controller
         $topic->save();
 
         return redirect()->route('forum_topic', ['slug' => $topic->slug, 'id' => $topic->id])
-            ->with($this->toastr->info('Label Change Has Been Applied', 'Info', ['options']));
+            ->withInfo('Label Change Has Been Applied');
     }
 
     public function suggestionTopic($slug, $id)
@@ -1045,7 +1038,7 @@ class ForumController extends Controller
         $topic->save();
 
         return redirect()->route('forum_topic', ['slug' => $topic->slug, 'id' => $topic->id])
-            ->with($this->toastr->info('Label Change Has Been Applied', 'Info', ['options']));
+            ->withInfo('Label Change Has Been Applied');
     }
 
     public function implementedTopic($slug, $id)
@@ -1059,6 +1052,6 @@ class ForumController extends Controller
         $topic->save();
 
         return redirect()->route('forum_topic', ['slug' => $topic->slug, 'id' => $topic->id])
-            ->with($this->toastr->info('Label Change Has Been Applied', 'Info', ['options']));
+            ->withInfo('Label Change Has Been Applied');
     }
 }
